+++ modules/hotpepper/optimize.go
@@ -1,0 +1,9 @@
+package hotpepper
+
+func Optimize() (*Condition, error) {
+	// todo: 位置情報とか食べたものから、hotpepper用のconditionを作って返す
+	return &Condition{}, nil
+}
+
+type Condition struct {
+}
